% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tracePlot}
\alias{tracePlot}
\title{Traceplots for Selected Parameters}
\usage{
tracePlot(x, parameter = "beta", index = 1)
}
\arguments{
\item{x}{The output object of \code{ULLGM_BMA}.}

\item{parameter}{Specifies which parameter should be considered for the traceplot. Options are \code{"beta"} for coefficients, \code{"alpha"} for the intercept (the default), \code{"modelsize"} for model size, and \code{"sigma2"} for the error variance.}

\item{index}{If \code{parameter = "beta"}, specifies which coefficient should be shown. Defaults to 1, corresponding to the covariate in the first column of X.}
}
\value{
Returns a 'ggplot2::ggplot' object.
}
\description{
\code{tracePlot} produces traceplots for selected parameters, extracted from \code{ULLGM_BMA} results.
}
\author{
Gregor Zens
}
